requirements.txt
streamlit==1.35.0
langchain==0.1.17
sentence-transformers==2.6.1
python-dotenv==1.0.1
google-generativeai==0.5.4
langchain-community==0.1.17